Add Latch in preference of Future
Rename and rework `Future` into `Latch` as `Latch` better describes
the behavior of the primitive.

`Latch` is a synchronization primitive that can be used to block until
a desired state is reached. The zero value for Latch is in an open
(blocking) state. Use the package level Resolve function to resolve
a Latch. Once resolved, the Latch cannot be reopened.